pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "notebooklens"
version = "0.2.0"
description = "AI-powered notebook diff review tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "NotebookLens Contributors"}
]
keywords = ["jupyter", "notebook", "github", "review", "diff", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "anthropic>=0.7.0",
    "pydantic>=2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
managed = [
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "PyJWT[crypto]>=2.9.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
]

dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "PyJWT[crypto]>=2.9.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
]

test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "PyJWT[crypto]>=2.9.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
